Change amount type to float64

+++ repository/account.go
@@ -1,12 +1,12 @@
 package repository
 
 type Account struct {
-	AccountID   int    `db:"account_id"`
-	CustomerID  int    `db:"customer_id"`
-	OpeningDate string `db:"opening_date"`
-	AccountType string `db:"account_type"`
-	Amount      int    `db:"amount"`
-	Status      int    `db:"status"`
+	AccountID   int     `db:"account_id"`
+	CustomerID  int     `db:"customer_id"`
+	OpeningDate string  `db:"opening_date"`
+	AccountType string  `db:"account_type"`
+	Amount      float64 `db:"amount"`
+	Status      int     `db:"status"`
 }
 
 type AccountRepository interface {
